fix new STLIB::Start and add vscode

+++ Core/Src/Examples/protection_manager.hpp
@@ -13,7 +13,7 @@ void protection_manager_example(void)
 	#define BOARD_ID 54
 	ProtectionManager::set_id(BOARD_ID);
 
-	double variable = 5.3;
+	int variable = 5.3;
 	ProtectionManager_add_protection(&variable, OUT_OF_RANGE, 1, 20, NOT_EQUALS, 5);
 
 	variable = 8;
